+++ CHANGELOG.md
@@ -21,7 +21,8 @@ but no memory unsafety
 
 Unreleased
 ==========
-* `sort_unstable*` variants
+* Added `sort_unstable*` variants
 * #![no_std] support with new opt-out feature "std"
   Stable sorts and OrdSubset implementations for f32 and f64 are unavailable in no_std mode 
   The latter can be regained but this depends on the unstable "core_float" feature gate (as of 2017-09-20, rustc 1.22), opt in via this crate's "unstable" feature
+* Generalized traits so that `.ord_subset_binary_search*` works on immutable slices and anything else that implements AsRef<[T]>, but not AsMut<[T]>